move ui header. updated CMakeLists.txt. still contains bugs.
