+++ .gitignore
@@ -7,6 +7,7 @@
 src/main.cpp
 res/*
 build/*
+ui/*
 =======
 build/*
 res/*
