+++ include/errorMsgHandler.h
@@ -4,7 +4,8 @@ Project:		TBD
 File:			errorMsgHandler.h
 Description:	Provides error messaging and handling
 */
-
+#ifndef _ERROR_MSG_HANDLE
+#define _ERROR_MSG_HANDLE
 namespace vba {
 	/*
 	file is null
@@ -15,5 +16,7 @@ namespace vba {
 	handler for empty file pointer
 	*/
 	void handleEmptyFilePtr (const char* const filename);
-
+	
 }
+
+#endif
