+++ src/errorMsgHandler.cpp
@@ -5,8 +5,7 @@ File:			errorMsgHandler.cpp
 Description:	Provides error messaging and handling
 */
 
-#ifndef _ERROR_MSG_HANDLE
-#define _ERROR_MSG_HANDLE
+
 
 #include <cstdlib>
 #include <iostream>
@@ -26,6 +25,7 @@ namespace vba {
 		std::cerr << "\nError opening " << filename << '\n';
 		exit(EXIT_FAILURE);
 	}
+
 };
 
-#endif
+
