+++ src/oni-to-pcd.cpp
@@ -5,8 +5,8 @@ File:			oni-to-pcd.cpp
 Description:	Reads an oni file recorded using the Openni2 or Openni library and outputs point clouds (pcd files)
 */
 
-#include "oni-to-pcd.h"
-#include "errorMsgHandler.h"
+#include "../include/oni-to-pcd.h"
+#include "../include/errorMsgHandler.h"
 #include <pcl/point_types.h>
 #include <pcl/io/pcd_io.h>
 #include <iostream>
@@ -22,10 +22,12 @@ Description:	Reads an oni file recorded using the Openni2 or Openni library and 
 here for execution of code in standalone, will be removed once 
 integrated in project
 ***************************************************************/
+/*
 int main (int argc, char* argv[]) {
 	/*
 	expect 2 additional arguments, ie. <executable> <oniFile> <pcdWriteDir>
 	*/
+/*
 	const int REQ_ARGS = 3;
 	if (argc < REQ_ARGS) {
  		std::cout << "\nNot enough arguments provided.\n";
@@ -50,11 +52,11 @@ int main (int argc, char* argv[]) {
 /***************************************************************
 vba::oni2pcd namespace functions
 ***************************************************************/
-
 /*
 read the given oni file, write the output pcd files to the given directory,
 if no directory is given will write to pcdTemp in the executable directory
 */
+
 void vba::oni2pcd::readOni (const char* oniFile, 
 		char* writeToDirPath,
 		const int framesToSkip) {
@@ -169,3 +171,4 @@ void vba::oni2pcd::writeCloudCb (const CloudConstPtr& cloud) {
 
 	++vba::oni2pcd::currentFrame;
 }
+
